+++ NEWS.md
@@ -5,8 +5,9 @@
 ### New APIs versions
 * Provides `indices v0.7`
 * Provides `search v1.3`
-* Requires `locations v3.0`
 * Provides `consortium-search v1.2`
+* Requires `locations v3.0`
+* Requires `instance-storage v10.3`
 
 ### Features
 * Implement indices recreation of the Linked Data resources ([MSEARCH-820](https://issues.folio.org/browse/MSEARCH-820))
@@ -38,6 +39,7 @@
 * Implement Reindexing of Libraries ([MSEARCH-766](https://issues.folio.org/browse/MSEARCH-766))
 * Implement Reindexing of Institutions ([MSEARCH-768](https://issues.folio.org/browse/MSEARCH-768))
 * Create computed field for sorting and filtering Date 1 ([MSEARCH-806](https://folio-org.atlassian.net/browse/MSEARCH-806))
+* Support filters for subject source and type on subject browse ([MSEARCH-805](https://folio-org.atlassian.net/browse/MSEARCH-805))
 
 ### Bug fixes
 * Do not delete kafka topics if collection topic is enabled ([MSEARCH-725](https://folio-org.atlassian.net/browse/MSEARCH-725))